Taking a way odd approach. Asset info generation project makes a source and header file with all info that can be pre-determined before compile time. Use it to load some shaders from the disk.

+++ common/math/maths.h
@@ -177,7 +177,7 @@ inline float Flicker(const char* pszValues, float flTime, float flLength)
 	if (!pszValues)
 		return 0;
 
-	int iValues = strlen(pszValues);
+	size_t iValues = strlen(pszValues);
 	if (iValues == 0)
 		return 0;
